swapped logic of the rubber treetap and rubber treelogs to allow for stripped rubber tree logs to be used as decoration blocks
